+++ account_tax_unece/__manifest__.py
@@ -1,14 +1,16 @@
-# Copyright 2016 Akretion (http://www.akretion.com)
+# Copyright 2016-2020 Akretion France (http://www.akretion.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 # @author Alexis de Lattre <alexis.delattre@akretion.com>
 
 {
     "name": "Account Tax UNECE",
-    "version": "13.0.1.0.0",
+    "version": "14.0.1.0.0",
     "category": "Accounting & Finance",
     "license": "AGPL-3",
+    "development_status": "Production/Stable",
     "summary": "UNECE nomenclature for taxes",
     "author": "Akretion,Odoo Community Association (OCA)",
+    "maintainers": ["alexis-via"],
     "website": "https://github.com/OCA/community-data-files",
     "depends": ["account", "base_unece"],
     "data": [
@@ -16,7 +18,6 @@
         "views/account_tax_template.xml",
         "data/unece_tax_type.xml",
         "data/unece_tax_categ.xml",
-        "data/unece_date.xml",
     ],
     "installable": True,
 }
