+++ account_tax_unece/models/unece_code_list.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2017 Akretion (http://www.akretion.com)
+# Copyright 2016-2020 Akretion France (http://www.akretion.com)
 # @author: Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
@@ -12,6 +12,9 @@ class UneceCodeList(models.Model):
         selection_add=[
             ("tax_type", "Tax Types (UNCL 5153)"),
             ("tax_categ", "Tax Categories (UNCL 5305)"),
-            ("date", "Date, Time or Period Qualifier (UNTDID 2005)"),
-        ]
+        ],
+        ondelete={
+            "tax_type": "cascade",
+            "tax_categ": "cascade",
+        },
     )
