+++ account_tax_unece/tests/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import test_account_invoice
+from . import test_tax_unece
