+++ account_tax_unece/views/account_tax.xml
@@ -1,24 +1,27 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <!--
-  Copyright 2016-2017 Akretion (http://www.akretion.com/)
+  Copyright 2016-2020 Akretion (http://www.akretion.com/)
   @author Alexis de Lattre <alexis.delattre@akretion.com>
   License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 <odoo>
-
-
-<record id="view_tax_form" model="ir.ui.view">
-    <field name="name">unece.account.tax.form</field>
-    <field name="model">account.tax</field>
-    <field name="inherit_id" ref="account.view_tax_form" />
-    <field name="arch" type="xml">
-        <group name="advanced_booleans" position="inside">
-            <field name="unece_type_id" context="{'default_type': 'tax_type'}" />
-            <field name="unece_categ_id" context="{'default_type': 'tax_categ'}" />
-            <field name="unece_due_date_id" context="{'default_type': 'date'}" />
-        </group>
-    </field>
-</record>
-
-
+    <record id="view_tax_form" model="ir.ui.view">
+        <field name="name">unece.account.tax.form</field>
+        <field name="model">account.tax</field>
+        <field name="inherit_id" ref="account.view_tax_form" />
+        <field name="arch" type="xml">
+            <group name="advanced_booleans" position="after">
+                <group name="unece">
+                    <field
+                        name="unece_type_id"
+                        context="{'default_type': 'tax_type'}"
+                    />
+                    <field
+                        name="unece_categ_id"
+                        context="{'default_type': 'tax_categ'}"
+                    />
+                </group>
+            </group>
+        </field>
+    </record>
 </odoo>
