+++ account_tax_unece/views/account_tax_template.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2017 Onestein (<http://www.onestein.eu>)
+<!-- Copyright 2017-2020 Onestein (<http://www.onestein.eu>)
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 <odoo>
-
     <record id="view_tax_template_form" model="ir.ui.view">
         <field name="model">account.tax.template</field>
         <field name="inherit_id" ref="account.view_account_tax_template_form" />
@@ -10,17 +9,17 @@
             <notebook position="inside">
                 <page string="Unece">
                     <group name="unece">
-                        <group>
-                            <field name="unece_type_id" />
-                            <field name="unece_categ_id" />
-                            <field name="unece_due_date_id" />
-                        </group>
-                        <group>
-                        </group>
+                        <field
+                            name="unece_type_id"
+                            options="{'no_create': True, 'no_create_edit': True}"
+                        />
+                        <field
+                            name="unece_categ_id"
+                            options="{'no_create': True, 'no_create_edit': True}"
+                        />
                     </group>
                 </page>
             </notebook>
         </field>
     </record>
-
 </odoo>
